Medical_Chatbot/requirement.txt
streamlit==1.31.0
langchain>=0.1.0
langchain-community>=0.0.17
langchain-core>=0.3.15
langchain-huggingface>=0.1.2
sentence-transformers==2.6.0
requests==2.32.2
pydantic>=2.9.0  # Updated to meet ollama's requirement
pinecone-client==3.0.0
python-dotenv==1.0.0
nltk==3.8.1
beautifulsoup4==4.12.2
ollama==0.4.4
datasets==3.2.0
langchain-ollama==0.2.2